backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
flask-jwt-extended>=4.4.0
flask-socketio>=5.3.6
psycopg2-binary>=2.9.9
pyjwt>=2.8.0
python-dotenv>=1.0.1
requests>=2.31.0
eventlet>=0.33.3